refactor SocketConnector
