+++ src/components/domain/Socket/SocketConnector.tsx
